Add collected entity to inventory UI.
